[ObjC]fix 64 to 32 bit clang conversion warning in src/core/lib (#29877)

+++ src/core/ext/transport/chttp2/transport/flow_control.h
@@ -259,7 +259,7 @@ class TransportFlowControl final {
   BdpEstimator* bdp_estimator() { return &bdp_estimator_; }
 
   uint32_t acked_init_window() const { return acked_init_window_; }
-  uint32_t sent_init_window() const { return target_initial_window_size_; }
+  int64_t sent_init_window() const { return target_initial_window_size_; }
 
   FlowControlAction SetAckedInitialWindow(uint32_t value);
